split policy_approximator apart, placing StateRepresentation in discretizer.py, add policy_representation.py
